refactor: Move Product/Project/Component logic from sboms to core app
This commit restructures the application by moving all Product, Project,
and Component management functionality from the sboms Django app to the
core app, while preserving the data models in sboms for backward
compatibility and to avoid complex migrations.

Changes include:
- Move Vue.js components and templates to core app
- Relocate admin registrations and API schemas
- Transfer views, URLs, and JavaScript logic to core
- Add legacy model comments in sboms/models.py
- Implement URL redirects from sboms to core endpoints
- Update import statements and component registrations

The sboms app now focuses exclusively on SBOM-specific functionality
while core handles the broader product lifecycle management. Data
models remain in sboms to preserve existing database structure and
avoid migration overhead.

This separation improves code organization and establishes clearer
boundaries between SBOM processing and general product management
features.
