+++ core/js/components/AddComponentForm.vue
