+++ core/js/components/AddProductForm.vue
