+++ core/js/components/ComponentMetaInfo.spec.ts
