+++ core/js/components/ComponentMetaInfo.vue
