+++ core/js/components/ComponentMetaInfoDisplay.vue
