+++ core/js/components/ComponentMetaInfoEditor.spec.ts
