+++ core/js/components/ComponentMetaInfoEditor.vue
