+++ core/js/components/DangerZone.spec.ts
