+++ core/js/components/DangerZone.vue
