+++ core/js/components/ItemAssignmentManager.spec.ts
