+++ core/js/components/ItemAssignmentManager.vue
