+++ core/js/components/ItemSelectModal.vue
