+++ core/js/components/ProductDangerZone.spec.ts
