+++ core/js/components/ProductDangerZone.vue
