+++ core/js/components/ProjectDangerZone.vue
