+++ core/js/enums.ts
@@ -1,0 +1,9 @@
+export enum LifecyclePhase {
+  Design = "design",
+  PreBuild = "pre-build",
+  Build = "build",
+  PostBuild = "post-build",
+  Operations = "operations",
+  Discovery = "discovery",
+  Decommission = "decommission"
+}
