+++ core/templates/core/component_details_private.html
