+++ core/templates/core/component_details_public.html
