+++ core/templates/core/components_dashboard.html
