+++ core/templates/core/project_details_private.html
