+++ core/templates/core/projects_dashboard.html
