+++ core/tests/test_crud_apis.py
@@ -16,15 +16,15 @@ from core.tests.fixtures import sample_user  # noqa: F401
 from teams.fixtures import sample_team_with_owner_member  # noqa: F401
 from teams.models import Member
 
-from ..models import Component, Product, Project
-from .fixtures import (  # noqa: F401
+from core.models import Component, Product, Project
+from sboms.tests.fixtures import (  # noqa: F401
     sample_access_token,
     sample_billing_plan,
     sample_component,
     sample_product,
     sample_project,
 )
-from .test_views import setup_test_session
+from sboms.tests.test_views import setup_test_session
 
 # =============================================================================
 # PRODUCT CRUD TESTS
