+++ core/urls.py
@@ -1,4 +1,5 @@
 from django.urls import path
+from django.views.generic import RedirectView
 
 from . import views
 
@@ -13,4 +14,78 @@ urlpatterns = [
     # Webhook support for Keycloak can be added here in the future if needed.
     # https://github.com/sbomify/sbomify/issues/69
     path("login", views.keycloak_login, name="keycloak_login"),
+    # Product/Project/Component URLs - moved from sboms app
+    # Backward compatibility redirects for URLs that now have trailing slashes
+    path("products", RedirectView.as_view(pattern_name="core:products_dashboard", permanent=True)),
+    path("projects", RedirectView.as_view(pattern_name="core:projects_dashboard", permanent=True)),
+    path("components", RedirectView.as_view(pattern_name="core:components_dashboard", permanent=True)),
+    path("product/<str:product_id>", RedirectView.as_view(pattern_name="core:product_details", permanent=True)),
+    path("project/<str:project_id>", RedirectView.as_view(pattern_name="core:project_details", permanent=True)),
+    path("component/<str:component_id>", RedirectView.as_view(pattern_name="core:component_details", permanent=True)),
+    path(
+        "public/product/<str:product_id>",
+        RedirectView.as_view(pattern_name="core:product_details_public", permanent=True),
+    ),
+    path(
+        "public/project/<str:project_id>",
+        RedirectView.as_view(pattern_name="core:project_details_public", permanent=True),
+    ),
+    path(
+        "public/component/<str:component_id>",
+        RedirectView.as_view(pattern_name="core:component_details_public", permanent=True),
+    ),
+    # Main URLs
+    path("products/", views.products_dashboard, name="products_dashboard"),
+    path(
+        "product/<str:product_id>/",
+        views.product_details_private,
+        name="product_details",
+    ),
+    path(
+        "public/product/<str:product_id>/",
+        views.product_details_public,
+        name="product_details_public",
+    ),
+    path("projects/", views.projects_dashboard, name="projects_dashboard"),
+    path(
+        "project/<str:project_id>/",
+        views.project_details_private,
+        name="project_details",
+    ),
+    path(
+        "public/project/<str:project_id>/",
+        views.project_details_public,
+        name="project_details_public",
+    ),
+    path("components/", views.components_dashboard, name="components_dashboard"),
+    path(
+        "component/<str:component_id>/",
+        views.component_details_private,
+        name="component_details",
+    ),
+    path(
+        "public/component/<str:component_id>/",
+        views.component_details_public,
+        name="component_details_public",
+    ),
+    path(
+        "component/<str:component_id>/transfer",
+        views.transfer_component_to_team,
+        name="transfer_component",
+    ),
+    path(
+        "project/<str:project_id>/sbom/download",
+        views.sbom_download_project,
+        name="sbom_download_project",
+    ),
+    path(
+        "product/<str:product_id>/sbom/download",
+        views.sbom_download_product,
+        name="sbom_download_product",
+    ),
+    path(
+        "component/<str:component_id>/metadata",
+        views.get_component_metadata,
+        name="get_component_metadata",
+    ),
 ]
