+++ sboms/js/components/ContactsEditor.vue
@@ -79,7 +79,7 @@
 
 <script setup lang="ts">
 import { ref } from 'vue';
-import type { ContactInfo } from '../type_defs';
+import type { ContactInfo } from '../../../core/js/type_defs.d.ts';
 
 interface Props {
   modelValue: ContactInfo[];
